@@ -18,22 +18,22 @@ export let results: RawResult[]=[
 {"f":"angular-v8.2.14-keyed","b":"31_startup-ci","v":[2896.3695000000007,2886.9935,2971.4145,2964.0045]},
 {"f":"angular-v8.2.14-keyed","b":"32_startup-bt","v":[199.764,186.844,208.236,193.37600000000003]},
 {"f":"angular-v8.2.14-keyed","b":"34_startup-totalbytes","v":[302.0537109375,302.0537109375,302.0537109375,302.0537109375]},
-{"f":"angular-ng-v10.0.4-keyed","b":"01_run1k","v":[172.241,174.395,208.404,186.238,171.045,173.894,175.391,165.328,171.434,165.66]},
-{"f":"angular-ng-v10.0.4-keyed","b":"02_replace1k","v":[139.225,136.432,140.036,135.224,136.122,137.328,136.969,142.164,136.798,138.276]},
-{"f":"angular-ng-v10.0.4-keyed","b":"03_update10th1k_x16","v":[198.046,163.63,172.717,175.601,158.996,168.401,180.592,164.593,167.556,163.86]},
-{"f":"angular-ng-v10.0.4-keyed","b":"04_select1k","v":[89.314,76.171,74.39,76.205,80.302,75.593,78.775,106.704,81.852,71.676]},
-{"f":"angular-ng-v10.0.4-keyed","b":"05_swap1k","v":[447.43,451.412,444.566,449.639,455.265,448.378,447.149,450.116,446.398,447.106]},
-{"f":"angular-ng-v10.0.4-keyed","b":"06_remove-one-1k","v":[24.089,27.231,27.375,27.554,28.416,27.864,27.577,29.471,27.716,30.077]},
-{"f":"angular-ng-v10.0.4-keyed","b":"07_create10k","v":[1321.716,1357.428,1407.647,1484.519,1536.848,1491.927,1446.23,1458.076,1376.979,1382.439]},
-{"f":"angular-ng-v10.0.4-keyed","b":"08_create1k-after1k_x2","v":[313.139,330.781,324.103,329.445,332.267,332.479,330.921,331.725,339.897,332.358]},
-{"f":"angular-ng-v10.0.4-keyed","b":"09_clear1k_x8","v":[260.902,258.58,260.717,263.065,322.278,268.119,270.865,270.688,274.137,254.772]},
-{"f":"angular-ng-v10.0.4-keyed","b":"21_ready-memory","v":[1.8780288696289062]},
-{"f":"angular-ng-v10.0.4-keyed","b":"22_run-memory","v":[4.305335998535156]},
-{"f":"angular-ng-v10.0.4-keyed","b":"23_update5-memory","v":[4.647068023681641]},
-{"f":"angular-ng-v10.0.4-keyed","b":"24_run5-memory","v":[5.140388488769531]},
-{"f":"angular-ng-v10.0.4-keyed","b":"25_run-clear-memory","v":[3.414867401123047]},
-{"f":"angular-ng-v10.0.4-keyed","b":"31_startup-ci","v":[2472.614,2449.73,2441.96,2476.21]},
-{"f":"angular-ng-v10.0.4-keyed","b":"32_startup-bt","v":[182.888,126.79200000000003,128.172,183.21200000000002]},
+{"f":"angular-ng-v10.0.4-keyed","b":"01_run1k","v":[172.951,172.09,172.114,165.124,175.823,174.397,178.443,171.187,170.621,166.036]},
+{"f":"angular-ng-v10.0.4-keyed","b":"02_replace1k","v":[137.267,139.215,139.806,139.432,141.981,138.192,139.948,139.276,138.417,141.008]},
+{"f":"angular-ng-v10.0.4-keyed","b":"03_update10th1k_x16","v":[170.799,170.664,177.139,171.952,177.704,182.797,191.824,181.992,167.85,167.337]},
+{"f":"angular-ng-v10.0.4-keyed","b":"04_select1k","v":[88.92,75.907,78.87,81.906,82.535,80.228,100.719,90.64,76.653,110.884]},
+{"f":"angular-ng-v10.0.4-keyed","b":"05_swap1k","v":[455.097,457.16,451.124,455.201,461.108,450.745,451.898,457.456,457.203,448.647]},
+{"f":"angular-ng-v10.0.4-keyed","b":"06_remove-one-1k","v":[24.016,35.516,27.193,31.32,29.324,29.336,30.114,29.707,27.365,27.822]},
+{"f":"angular-ng-v10.0.4-keyed","b":"07_create10k","v":[1328.046,1376.638,1407.63,1505.027,1548.988,1510.109,1488.084,1470.989,1402.185,1417.741]},
+{"f":"angular-ng-v10.0.4-keyed","b":"08_create1k-after1k_x2","v":[312.053,330.574,328.473,332.409,331.87,328.56,326.056,325.976,329.651,336.938]},
+{"f":"angular-ng-v10.0.4-keyed","b":"09_clear1k_x8","v":[258.907,267.437,262.067,263.102,268.39,264.405,273.575,265.451,270.946,268.587]},
+{"f":"angular-ng-v10.0.4-keyed","b":"21_ready-memory","v":[1.878082275390625]},
+{"f":"angular-ng-v10.0.4-keyed","b":"22_run-memory","v":[4.305366516113281]},
+{"f":"angular-ng-v10.0.4-keyed","b":"23_update5-memory","v":[4.647083282470703]},
+{"f":"angular-ng-v10.0.4-keyed","b":"24_run5-memory","v":[5.140357971191406]},
+{"f":"angular-ng-v10.0.4-keyed","b":"25_run-clear-memory","v":[3.41485595703125]},
+{"f":"angular-ng-v10.0.4-keyed","b":"31_startup-ci","v":[2440.2019999999998,2439.69,2439.25,2439.146]},
+{"f":"angular-ng-v10.0.4-keyed","b":"32_startup-bt","v":[133.31999999999996,133.11999999999998,133.99600000000004,128.748]},
 {"f":"angular-ng-v10.0.4-keyed","b":"34_startup-totalbytes","v":[295.3486328125,295.3486328125,295.3486328125,295.3486328125]},
 {"f":"angular-noopzone-v8.0.1-keyed","b":"01_run1k","v":[143.23,139.607,145.329,140.737,139.714,146.385,150.683,135.616,135.255,154.326]},
 {"f":"angular-noopzone-v8.0.1-keyed","b":"02_replace1k","v":[133.947,129.618,134.842,131.124,129.708,132.001,130.679,135.117,138.574,131.384]},
@@ -1106,23 +1106,23 @@ export let results: RawResult[]=[
 {"f":"react-easy-state-v16.8.6 + 6.3.0-keyed","b":"31_startup-ci","v":[2607.3599999999997,2608.33,2691.38,2610.29]},
 {"f":"react-easy-state-v16.8.6 + 6.3.0-keyed","b":"32_startup-bt","v":[53.511999999999986,57.76800000000001,102.93200000000004,59.84400000000002]},
 {"f":"react-easy-state-v16.8.6 + 6.3.0-keyed","b":"34_startup-totalbytes","v":[270.1357421875,270.1357421875,270.1357421875,270.1357421875]},
-{"f":"react-hooks-v16.8.6-keyed","b":"01_run1k","v":[164.755,160.1,162.025,159.713,164.001,159.797,165.147,160.254,169.127,158.024]},
-{"f":"react-hooks-v16.8.6-keyed","b":"02_replace1k","v":[131.656,132.368,131.031,132.877,133.97,134.647,133.711,133.239,133.818,130.953]},
-{"f":"react-hooks-v16.8.6-keyed","b":"03_update10th1k_x16","v":[180.004,187.791,184.633,174.189,182.433,189.921,177.492,180.851,178.022,181.005]},
-{"f":"react-hooks-v16.8.6-keyed","b":"04_select1k","v":[80.006,88.776,96.168,65.726,98.264,85.139,94.689,71.088,68.372,91.598]},
-{"f":"react-hooks-v16.8.6-keyed","b":"05_swap1k","v":[431.164,440.156,440.944,445.295,437.823,438.18,441.2,438.306,443.468,433.016]},
-{"f":"react-hooks-v16.8.6-keyed","b":"06_remove-one-1k","v":[25.92,28.456,24.522,24.5,25.802,24.669,24.459,24.987,24.971,36.291]},
-{"f":"react-hooks-v16.8.6-keyed","b":"07_create10k","v":[1509.111,1684.751,1681.44,1705.623,1721.389,1732.36,1733.604,1727.876,1681.677,1698.379]},
-{"f":"react-hooks-v16.8.6-keyed","b":"08_create1k-after1k_x2","v":[303.456,309.072,313.25,308.499,308.972,312.849,315.171,318.252,307.946,317.39]},
-{"f":"react-hooks-v16.8.6-keyed","b":"09_clear1k_x8","v":[145.59,140.279,138.341,140.962,146.121,140.317,141.377,138.566,141.242,139.858]},
+{"f":"react-hooks-v16.8.6-keyed","b":"01_run1k","v":[166.288,164.231,162.579,161.213,174.123,158.68,164.884,159.285,166.878,158.298]},
+{"f":"react-hooks-v16.8.6-keyed","b":"02_replace1k","v":[131.684,134.882,132.756,131.356,128.009,135.231,133.646,132.666,134.867,130.309]},
+{"f":"react-hooks-v16.8.6-keyed","b":"03_update10th1k_x16","v":[202.431,191.248,180.948,179.376,180.537,188.747,200.117,196.285,181.345,184.265]},
+{"f":"react-hooks-v16.8.6-keyed","b":"04_select1k","v":[78.868,91.536,94.345,66.597,98.002,92.283,95.111,68.548,91.138,94.24]},
+{"f":"react-hooks-v16.8.6-keyed","b":"05_swap1k","v":[432.697,442.072,440.686,452.615,443.522,450.472,451.095,438.087,437.616,466.385]},
+{"f":"react-hooks-v16.8.6-keyed","b":"06_remove-one-1k","v":[24.977,24.244,24.435,24.512,26.898,33.71,25.006,26.409,24.455,24.312]},
+{"f":"react-hooks-v16.8.6-keyed","b":"07_create10k","v":[1510.787,1719.75,1670.153,1729.554,1580.351,1775.911,1736.399,1754.262,1746.433,1712.006]},
+{"f":"react-hooks-v16.8.6-keyed","b":"08_create1k-after1k_x2","v":[306.802,315.015,310.269,313.736,308.985,312.35,320.46,317.628,310.676,330.502]},
+{"f":"react-hooks-v16.8.6-keyed","b":"09_clear1k_x8","v":[147.125,142.502,142.729,143.051,144.505,142.369,148.365,143.811,144.111,143.282]},
 {"f":"react-hooks-v16.8.6-keyed","b":"21_ready-memory","v":[1.3639869689941406]},
-{"f":"react-hooks-v16.8.6-keyed","b":"22_run-memory","v":[4.0996856689453125]},
-{"f":"react-hooks-v16.8.6-keyed","b":"23_update5-memory","v":[4.877464294433594]},
-{"f":"react-hooks-v16.8.6-keyed","b":"24_run5-memory","v":[5.17303466796875]},
-{"f":"react-hooks-v16.8.6-keyed","b":"25_run-clear-memory","v":[3.1529083251953125]},
-{"f":"react-hooks-v16.8.6-keyed","b":"31_startup-ci","v":[2694.77,2578.34,2596.7599999999998,2583.23]},
-{"f":"react-hooks-v16.8.6-keyed","b":"32_startup-bt","v":[109.192,16,70.304,16]},
-{"f":"react-hooks-v16.8.6-keyed","b":"34_startup-totalbytes","v":[260.0107421875,260.0107421875,260.0107421875,260.0107421875]},
+{"f":"react-hooks-v16.8.6-keyed","b":"22_run-memory","v":[4.099617004394531]},
+{"f":"react-hooks-v16.8.6-keyed","b":"23_update5-memory","v":[4.877346038818359]},
+{"f":"react-hooks-v16.8.6-keyed","b":"24_run5-memory","v":[5.176982879638672]},
+{"f":"react-hooks-v16.8.6-keyed","b":"25_run-clear-memory","v":[3.1527328491210938]},
+{"f":"react-hooks-v16.8.6-keyed","b":"31_startup-ci","v":[2581.165,2582.325,2581.23,2579.435]},
+{"f":"react-hooks-v16.8.6-keyed","b":"32_startup-bt","v":[16,16,16,16]},
+{"f":"react-hooks-v16.8.6-keyed","b":"34_startup-totalbytes","v":[260.015625,260.015625,260.015625,260.015625]},
 {"f":"react-mobX-v16.13.1 + 5.15.4-keyed","b":"01_run1k","v":[219.522,214.676,214.342,212.373,211.965,211.39,211.979,207.368,206.912,216.707]},
 {"f":"react-mobX-v16.13.1 + 5.15.4-keyed","b":"02_replace1k","v":[147.314,149.618,151.301,154.723,151.573,151.991,147.213,147.906,152.855,150.279]},
 {"f":"react-mobX-v16.13.1 + 5.15.4-keyed","b":"03_update10th1k_x16","v":[225.432,228.098,228.085,240.145,223.318,222.304,229.227,222.531,224.246,232.081]},
@@ -1514,21 +1514,21 @@ export let results: RawResult[]=[
 {"f":"uhtml-v1.8.1-keyed","b":"31_startup-ci","v":[1955.4399999999998,1956.005,1986.725,1954.555]},
 {"f":"uhtml-v1.8.1-keyed","b":"32_startup-bt","v":[16,16,16,16]},
 {"f":"uhtml-v1.8.1-keyed","b":"34_startup-totalbytes","v":[146.8076171875,146.8076171875,146.8076171875,146.8076171875]},
-{"f":"vanillajs-keyed","b":"01_run1k","v":[111.88,107.127,103.3,107.149,105.211,107.139,107.345,105.361,104.592,105.687]},
-{"f":"vanillajs-keyed","b":"02_replace1k","v":[113.53,106.271,109.136,107.905,108.948,107.434,116.883,107.973,108.662,108.381]},
-{"f":"vanillajs-keyed","b":"03_update10th1k_x16","v":[143.276,136.564,140.021,158.157,141.663,149.432,148.405,138.277,146.633,151.233]},
-{"f":"vanillajs-keyed","b":"04_select1k","v":[28.526,24.452,29.555,20.531,18.475,24.179,24.685,19.791,28.157,20.276]},
-{"f":"vanillajs-keyed","b":"05_swap1k","v":[46.943,46.481,46.355,45.984,50.101,46.615,47.158,47.374,47.447,46.448]},
-{"f":"vanillajs-keyed","b":"06_remove-one-1k","v":[23.489,22.905,26.053,23.391,26.984,31.671,23.343,23.129,23.145,26.962]},
-{"f":"vanillajs-keyed","b":"07_create10k","v":[999.268,990.917,1000.991,1029.196,1006.702,1037.654,1015.615,1036.045,1001.672,1038.385]},
-{"f":"vanillajs-keyed","b":"08_create1k-after1k_x2","v":[220.453,224.516,224.502,219.85,233.441,224.596,225.224,223.677,221.219,225.288]},
-{"f":"vanillajs-keyed","b":"09_clear1k_x8","v":[102.49,106.102,105.696,104.415,106.287,106.449,111.317,110.769,104.214,106.445]},
-{"f":"vanillajs-keyed","b":"21_ready-memory","v":[1.0822296142578125]},
-{"f":"vanillajs-keyed","b":"22_run-memory","v":[1.6160888671875]},
-{"f":"vanillajs-keyed","b":"23_update5-memory","v":[1.9403457641601562]},
-{"f":"vanillajs-keyed","b":"24_run5-memory","v":[2.2528038024902344]},
-{"f":"vanillajs-keyed","b":"25_run-clear-memory","v":[2.294281005859375]},
-{"f":"vanillajs-keyed","b":"31_startup-ci","v":[1955.0400000000002,1963.4899999999998,1964.02,1955.4499999999998]},
+{"f":"vanillajs-keyed","b":"01_run1k","v":[108.894,104.479,109.003,103.069,104.043,104.692,104.568,105.758,103.595,109.606]},
+{"f":"vanillajs-keyed","b":"02_replace1k","v":[108.083,110.045,108.773,110.099,109.041,108.968,106.876,107.865,108.017,108.683]},
+{"f":"vanillajs-keyed","b":"03_update10th1k_x16","v":[151.089,143.615,144.676,147.407,158.001,155.012,143.114,144.512,139.979,152.765]},
+{"f":"vanillajs-keyed","b":"04_select1k","v":[27.18,22.709,22.983,23.889,21.61,24.013,21.608,20.077,24.114,23.321]},
+{"f":"vanillajs-keyed","b":"05_swap1k","v":[46.602,49.069,52.817,48.127,47.201,50.328,48.705,48.174,49.182,49.646]},
+{"f":"vanillajs-keyed","b":"06_remove-one-1k","v":[25.971,23.341,23.44,35.494,26.613,23.55,23.548,23.483,23.671,26.37]},
+{"f":"vanillajs-keyed","b":"07_create10k","v":[995.323,994.694,1016.779,1035.801,1002.515,1047.409,995.246,1048.787,1012.826,1057.733]},
+{"f":"vanillajs-keyed","b":"08_create1k-after1k_x2","v":[219.553,225.451,217.236,221.008,223.603,220.319,225.908,224.613,224.512,224.153]},
+{"f":"vanillajs-keyed","b":"09_clear1k_x8","v":[104.905,109.688,106.694,108.028,110.037,108.122,106.216,110.725,104.777,106.383]},
+{"f":"vanillajs-keyed","b":"21_ready-memory","v":[1.082244873046875]},
+{"f":"vanillajs-keyed","b":"22_run-memory","v":[1.6160316467285156]},
+{"f":"vanillajs-keyed","b":"23_update5-memory","v":[1.9401092529296875]},
+{"f":"vanillajs-keyed","b":"24_run5-memory","v":[2.2526397705078125]},
+{"f":"vanillajs-keyed","b":"25_run-clear-memory","v":[2.2942581176757812]},
+{"f":"vanillajs-keyed","b":"31_startup-ci","v":[1955.275,1955.79,1954.89,1954.9]},
 {"f":"vanillajs-keyed","b":"32_startup-bt","v":[16,16,16,16]},
 {"f":"vanillajs-keyed","b":"34_startup-totalbytes","v":[150.125,150.125,150.125,150.125]},
 {"f":"vanillajs-1-keyed","b":"01_run1k","v":[115.645,104.25,103.902,106.799,105.097,103.517,106.946,106.458,109.074,102.885]},
